Pattern recognition back to working in similar state. The problem was that we pruned too many track in comparing if they had similar hits

+++ include/EUTelGeometryTelescopeGeoDescription.h
@@ -302,7 +302,7 @@ namespace eutelescope {
 
 		int findIntersectionWithCertainID( float x0, float y0, float z0, float px, float py, float pz, float _beamQ, int nextPlaneID, float output[]);
 
-		TVector3 getXYZfromArcLength( float x0, float y0, float z0, float px, float py, float pz, float _beamQ, float s) const;
+		TVector3 getXYZfromArcLength( double x0, double y0, double z0, double px, double py, double pz, float _beamQ, double s) const;
 
 		TMatrix getPropagationJacobianF( float x0, float y0, float z0, float px, float py, float pz, float _beamQ, float dz );
